+++ test/language/indexed_offsets/cpp/OptionalNestedIndexedOffsetArrayTest.cpp
@@ -26,6 +26,7 @@ public:
         for (uint8_t i = 0; i < NUM_ELEMENTS; ++i)
             m_data.push_back(data[i]);
     }
+
 protected:
     void writeOptionalNestedIndexedOffsetArrayToByteArray(zserio::BitStreamWriter& writer, uint16_t length,
             bool writeWrongOffsets)
@@ -36,7 +37,7 @@ protected:
         {
             const uint32_t wrongOffset = WRONG_OFFSET;
             uint32_t currentOffset = ELEMENT0_OFFSET;
-            for (uint8_t i = 0; i < length; ++i)
+            for (uint16_t i = 0; i < length; ++i)
             {
                 if ((i + 1) == length && writeWrongOffsets)
                     writer.writeBits(wrongOffset, 32);
@@ -46,7 +47,7 @@ protected:
             }
 
             // already aligned
-            for (uint8_t i = 0; i < length; ++i)
+            for (uint16_t i = 0; i < length; ++i)
                 writer.writeString(m_data[i]);
         }
 
@@ -60,7 +61,7 @@ protected:
         const vector_type<uint32_t>& offsets = optionalNestedIndexedOffsetArray.getHeader().getOffsets();
         ASSERT_EQ(length, offsets.size());
         uint32_t expectedOffset = ELEMENT0_OFFSET + offsetShift;
-        for (uint8_t i = 0; i < length; ++i)
+        for (uint16_t i = 0; i < length; ++i)
         {
             ASSERT_EQ(expectedOffset, offsets[i]);
             expectedOffset += static_cast<uint32_t>(zserio::bitSizeOfString(m_data[i]) / 8);
@@ -68,7 +69,7 @@ protected:
     }
 
     void checkOptionalNestedIndexedOffsetArray(
-            const OptionalNestedIndexedOffsetArray& optionalNestedIndexedOffsetArray, int16_t length)
+            const OptionalNestedIndexedOffsetArray& optionalNestedIndexedOffsetArray, uint16_t length)
     {
         ASSERT_EQ(length, optionalNestedIndexedOffsetArray.getHeader().getLength());
 
@@ -79,7 +80,7 @@ protected:
         {
             const vector_type<string_type>& data = optionalNestedIndexedOffsetArray.getData();
             ASSERT_EQ(length, data.size());
-            for (uint8_t i = 0; i < length; ++i)
+            for (uint16_t i = 0; i < length; ++i)
                 ASSERT_EQ(m_data[i], data[i]);
         }
 
@@ -97,7 +98,7 @@ protected:
         offsets.reserve(length);
         const uint32_t wrongOffset = WRONG_OFFSET;
         uint32_t currentOffset = ELEMENT0_OFFSET;
-        for (uint8_t i = 0; i < length; ++i)
+        for (uint16_t i = 0; i < length; ++i)
         {
             if ((i + 1) == length && createWrongOffsets)
                 offsets.push_back(wrongOffset);
